deployment/01.restaurant_assistant_fargate/agent-requirements.txt
boto3
botocore
awscli
opensearch-py
requests-aws4auth
pyyaml
retrying
strands-agents
strands-agents-tools